pyproject.toml
[tool.poetry]
name = "billybot"
version = "0.1.0"
description = ""
authors = ["luka"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.1.0"
python-dotenv = "^0.21.0"
tabulate = "^0.9.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
